ES Upload/requirements.txt
sentence-transformers==3.1.0
langchain==0.3.0
mysql-connector-python==9.0.0